requirements.txt
django==6.0.3
djangorestframework==3.16.1
gunicorn==25.1.0
psycopg2-binary==2.9.11

